# backend/restaurant-service/src/main/java/ru/isavelev76/restaurantservice/controllers/RestaurantController.java
package ru.isavelev76.restaurantservice.controllers;

import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.*;
import ru.isavelev76.restaurantservice.requests.RestaurantFilterRequest;
import ru.isavelev76.restaurantservice.requests.RestaurantRequest;
import ru.isavelev76.restaurantservice.responses.RestaurantResponse;
import ru.isavelev76.restaurantservice.services.RestaurantService;

import java.util.UUID;

/**
 * @author Ilya Savelyev
 * @since 09.12.2025
 */

@RestController
@RequestMapping("/api/restaurants")
@RequiredArgsConstructor
public class RestaurantController {

    private final RestaurantService restaurantService;

    @GetMapping()
    public ResponseEntity<Page<RestaurantResponse>> getAll(@ModelAttribute RestaurantFilterRequest filter,
                                                           Pageable pageable) {
        return ResponseEntity.ok(restaurantService.getFilteredRestaurants(filter, pageable));
    }

    @GetMapping("/{id}")
    public ResponseEntity<RestaurantResponse> getById(@PathVariable UUID id) {
        return ResponseEntity.ok(restaurantService.getById(id));
    }

    @PostMapping
    @SecurityRequirement(name = "JWT")
    @PreAuthorize("hasRole('ROLE_ADMIN')")
    public ResponseEntity<RestaurantResponse> create(@RequestBody RestaurantRequest request) {
        return ResponseEntity.status(HttpStatus.CREATED)
                .body(restaurantService.create(request));
    }

    @PutMapping("/{id}")
    @SecurityRequirement(name = "JWT")
    @PreAuthorize("hasRole('ROLE_ADMIN')")
    public ResponseEntity<RestaurantResponse> update(@PathVariable UUID id,
                                                     @RequestBody RestaurantRequest request) {
        return ResponseEntity.ok(restaurantService.update(id, request));
    }

    @DeleteMapping("/{id}")
    @SecurityRequirement(name = "JWT")
    @PreAuthorize("hasRole('ROLE_ADMIN')")
    public ResponseEntity<Void> delete(@PathVariable UUID id) {
        restaurantService.delete(id);
        return ResponseEntity.noContent().build();
    }
}