# backend/restaurant-service/src/main/resources/swagger-ui.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:9090",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/restaurants/{id}": {
            "get": {
                "tags": [
                    "restaurant-controller"
                ],
                "operationId": "getById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RestaurantResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "restaurant-controller"
                ],
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RestaurantRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RestaurantResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "restaurant-controller"
                ],
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/dishes/{id}": {
            "get": {
                "tags": [
                    "dish-controller"
                ],
                "operationId": "getById_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DishResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "dish-controller"
                ],
                "operationId": "updateDish",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "dish": {
                                        "$ref": "#/components/schemas/DishRequest"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "dish"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DishResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "dish-controller"
                ],
                "operationId": "delete_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/restaurants": {
            "get": {
                "tags": [
                    "restaurant-controller"
                ],
                "operationId": "getAll",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/RestaurantFilterRequest"
                        }
                    },
                    {
                        "name": "pageable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageRestaurantResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "restaurant-controller"
                ],
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RestaurantRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RestaurantResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/dishes": {
            "post": {
                "tags": [
                    "dish-controller"
                ],
                "operationId": "createDish",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "dish": {
                                        "$ref": "#/components/schemas/DishRequest"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "dish"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DishResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/dishes/prices": {
            "post": {
                "tags": [
                    "dish-controller"
                ],
                "operationId": "getDishPrices",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/DishPriceRequest"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DishPriceResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/dishes/restaurant/{restaurantId}": {
            "get": {
                "tags": [
                    "dish-controller"
                ],
                "operationId": "getAllByRestaurant",
                "parameters": [
                    {
                        "name": "restaurantId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DishResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/dishes/media/{fileName}": {
            "get": {
                "tags": [
                    "dish-controller"
                ],
                "operationId": "getImage",
                "parameters": [
                    {
                        "name": "fileName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "byte"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "RestaurantRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1
                    },
                    "cuisine": {
                        "type": "string",
                        "minLength": 1
                    },
                    "address": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "address",
                    "cuisine",
                    "name"
                ]
            },
            "RestaurantResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "cuisine": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    }
                }
            },
            "DishRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    },
                    "price": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    },
                    "restaurantId": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "name",
                    "price",
                    "restaurantId"
                ]
            },
            "DishResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "restaurantId": {
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
            "DishPriceRequest": {
                "type": "object",
                "properties": {
                    "dishId": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "dishId"
                ]
            },
            "DishPriceResponse": {
                "type": "object",
                "properties": {
                    "dishId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "price": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "RestaurantFilterRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "cuisine": {
                        "type": "string"
                    }
                }
            },
            "Pageable": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "PageRestaurantResponse": {
                "type": "object",
                "properties": {
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RestaurantResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sort": {
                        "$ref": "#/components/schemas/SortObject"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageable": {
                        "$ref": "#/components/schemas/PageableObject"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
            "PageableObject": {
                "type": "object",
                "properties": {
                    "offset": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "sort": {
                        "$ref": "#/components/schemas/SortObject"
                    },
                    "paged": {
                        "type": "boolean"
                    },
                    "pageNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "unpaged": {
                        "type": "boolean"
                    }
                }
            },
            "SortObject": {
                "type": "object",
                "properties": {
                    "empty": {
                        "type": "boolean"
                    },
                    "sorted": {
                        "type": "boolean"
                    },
                    "unsorted": {
                        "type": "boolean"
                    }
                }
            }
        }
    }
}